requirements.txt
fastapi==0.117.1
uvicorn[standard]==0.36.0
SQLAlchemy==2.0.43
pydantic==2.11.9
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
email-validator==2.3.0
bcrypt==4.0.1
python-dotenv==1.1.1
